Changed all scene changes with new method added.
