+++ src/scenes/Scene1,2forgotpasswordpopup.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.*?>
 
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.ChangeScene">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.Dashboard">
    <children>
       <Button id="BTchangepassword" fx:id="BTchangepassword" defaultButton="true" layoutX="99.0" layoutY="261.0" mnemonicParsing="false" onAction="#login" text="Change password" />
       <Label layoutX="99.0" layoutY="159.0" text="Confirm your password" />
