+++ src/scenes/Scene3,3,1deleteroomconfirmpopup.fxml
@@ -4,7 +4,7 @@
 <?import java.lang.*?>
 <?import javafx.scene.layout.*?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="200.0" prefWidth="199.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.ChangeScene">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="200.0" prefWidth="199.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.Dashboard">
    <children>
       <Label layoutX="4.0" layoutY="45.0" text="Please type in the room name again&#10;and press confirm" textAlignment="CENTER" />
       <TextField id="TFconfirmdeleteroom" fx:id="TFconfirmdeleteroom" layoutX="25.0" layoutY="90.0" />
