+++ src/scenes/Scene4,3deleteaccountpopup.fxml
@@ -4,7 +4,7 @@
 <?import java.lang.*?>
 <?import javafx.scene.layout.*?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="200.0" prefWidth="199.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.ChangeScene">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="200.0" prefWidth="199.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.Dashboard">
    <children>
       <Label layoutX="25.0" layoutY="38.0" text="Password" textAlignment="CENTER" />
       <Button id="BTconfirmdeleteaccount" fx:id="BTconfirmdeleteaccount" layoutX="31.0" layoutY="152.0" mnemonicParsing="false" onAction="#BackToLogin" text="Delete account forever&#10;(a very long time)" textAlignment="CENTER" />
