+++ src/scenes/Scene6,1ConfimBooking.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.text.Text?>
 
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.ChangeScene">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.Dashboard">
    <children>
       <Button id="BThelp" fx:id="BThelp" layoutX="54.0" layoutY="44.0" mnemonicParsing="false" text="Help" />
       <Text layoutX="234.0" layoutY="173.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Do you wish to book this room?" />
