+++ src/scenes/Scene8,1ConfirmDelete.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.text.Text?>
 
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.ChangeScene">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="scenes.Dashboard">
    <children>
       <Button id="BThelp" fx:id="BThelp" layoutX="34.0" layoutY="24.0" mnemonicParsing="false" text="Help" />
       <Text layoutX="205.0" layoutY="154.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Do you wish to delete this booking?" />
